Add a DependencyAPI to firrtl.options.Phase
This adds a DependencyAPI trait for Phase. This defines two member
functions that form the basis of defining dependencies between Phases:
prerequisites and invalidates. Prerequisites define a set of Phases
that must run *before* some Phase. Invalidates defines a function that
can be called to filter some collection of Phases.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>